Adding properties for base rest api, moving id to integer from long

+++ src/main/java/org/nhsrc/repository/BaseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import java.util.UUID;
 
 @NoRepositoryBean
-public interface BaseRepository<T> extends PagingAndSortingRepository<T, Long> {
+public interface BaseRepository<T> extends PagingAndSortingRepository<T, Integer> {
     T find(UUID uuid);
 }
